refactor(server): split create_app route handlers into domain route-modules (Tier-3 A2)
